Moved tests to proper file
