+++ dateparser/conf.py
@@ -38,6 +38,7 @@ class Settings(object):
     PREFER_DATES_FROM = 'current_period'  # past, future, current_period
     SUPPORT_BEFORE_COMMON_ERA = False
     PREFER_DAY_OF_MONTH = 'current'  # current, first, last
+    SKIP_TOKENS = ['t']
 
     def __init__(self, **kwargs):
         for key in kwargs:
